% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composerr.R
\name{composerr_get_action}
\alias{composerr_get_action}
\title{Retrieve the default ultimate error handler of a \link[=composerr]{composerr error handler}}
\usage{
composerr_get_action(err_h)
}
\arguments{
\item{err_h}{An error handler created with \code{\link[=composerr]{composerr()}}.}
}
\value{
The \strong{default ultimate error processing function}
\code{action} assigned to \code{err_h <- composerr(..., action = my_default_action)}.
If \code{err_h} is a child error handler of \code{err_h_parent}
has been created without assigning a default ultimate error handler \code{action},
then the default ultimate error handler assigned to \code{err_h_parent}
is returned by the call \code{composerr_get_action(err_h)}.
If \code{err_h_parent} also has no default ultimate error handler assigned to
it, then this process is repeated recursively until the first ancestor
of \code{err_h} is found that has
a non-null default ultimate error handler \code{action} assigned to.
If no default ultimate error handler was assigned for any error handler
in the entire error handler cascade, then \code{\link[=stop]{stop()}} will be the return
value of \code{composerr_get_action(err_h)}.
}
\description{
Retrieve the default ultimate error handler of a \link[=composerr]{composerr error handler}
}
\seealso{
\code{\link[=composerr]{composerr()}}, \code{\link[=composerr_flush]{composerr_flush()}}, \code{\link[=composerr_halt]{composerr_halt()}}
and \code{\link[=validate_composerr]{validate_composerr()}}
}
