% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composerr.R
\name{composerr_get_internal_handler}
\alias{composerr_get_internal_handler}
\title{Retrieve the internal error processing function of an \link[=composerr]{error handler}}
\usage{
composerr_get_internal_handler(err_h)
}
\arguments{
\item{err_h}{An error handler created with \code{\link[=composerr]{composerr()}}.}
}
\value{
The internal error processing function
\code{internal_handler()} assigned to \code{err_h()}.
If \code{err_h()} was created from another error handling function \code{err_h_old()}
and \code{err_h()} has been created without assigning a new \code{internal_handler},
then the \code{internal_handler()} assigned to \code{err_h_old()} is returned when
calling \code{composerr_get_handler(err_h)}. This process is repeated
recursively until the first ancestor of \code{err_h()} which has assigned an
\code{internal_handler()}.
}
\description{
Retrieve the internal error processing function of an \link[=composerr]{error handler}
}
\seealso{
\code{\link[=composerr]{composerr()}}, \code{\link[=composerr_modify]{composerr_modify()}}, \code{\link[=composerr_flush]{composerr_flush()}}, \code{\link[=composerr_halt]{composerr_halt()}}
and \code{\link[=validate_composerr]{validate_composerr()}}
}
