% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stringify}
\alias{stringify}
\title{Stringify vector}
\usage{
stringify(
  x,
  str_collapse = ", ",
  str_quote = "'",
  str_before = "",
  str_after = "",
  new_line = FALSE
)
}
\arguments{
\item{x}{An object, which should be turned into a string. Usually an atomic
object (a vector).}

\item{str_collapse}{A string, used for separating each vector entry.}

\item{str_quote}{A string, holding  the symbol which should be used for quoting
every entry of \code{x}. Default is \verb{'} which gives the result \code{'entry1'}.}

\item{str_before}{A string placed before each vector entry.}

\item{str_after}{A string placed after each vector entry.}

\item{new_line}{A logical. If set to \code{TRUE}, then a new line command is
placed at the end of the resulting string.}
}
\value{
A string showing the entries of the vector.
}
\description{
Turn a vector (e.g. \code{c(1, 2, 3)}) into a string (e.g. \code{"'1', '2', '3'"}).
This function is particularly useful when creating error messages,
describing problematic values.
}
