% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composerr.R
\name{validate_composerr}
\alias{validate_composerr}
\title{Validate an \link[=composerr]{error handler}}
\usage{
validate_composerr(obj, err_h, allow_null = TRUE, obj_name = NULL)
}
\arguments{
\item{obj}{An error handling function created by \code{composerr()}.}

\item{err_h}{An \link[=composerr]{error handler} called when
\code{obj} is not a valid error handling function created by \code{\link[=composerr]{composerr()}}.}

\item{allow_null}{A logical flag, defining if \code{obj = NULL} should be allowed.}

\item{obj_name}{An optional string, defining the variable name of
\code{obj}. This string is used for creating a meaningful error message,
in case the validation failed. If \code{obj_name} is omitted, then \code{obj_name}
is calculated by using \strong{non-standard-evaluation} on \code{obj}.}
}
\description{
Check that \code{obj} is really an error handler created with \code{\link[=composerr]{composerr()}}.
}
\seealso{
\code{\link[=composerr]{composerr()}}, \code{\link[=composerr_flush]{composerr_flush()}},
\code{\link[=composerr_halt]{composerr_halt()}}, \code{\link[=composerr_get_action]{composerr_get_action()}} and
\code{\link[=validate_composerr]{validate_composerr()}}
}
